Move parser into separate package folder
