+++ src/main/java/sylvie/storage/parser/DeadlineParser.java
@@ -1,22 +1,21 @@
-package sylvie.storage;
+package sylvie.storage.parser;
 
 import sylvie.exception.IllegalDataException;
 import sylvie.exception.InvalidArgumentException;
-import sylvie.task.Event;
+import sylvie.task.Deadline;
 import sylvie.task.Task;
 
-public class EventParser extends TaskParser {
+public class DeadlineParser extends TaskParser {
     @Override
     public Task parse(String input) throws IllegalDataException {
         String[] parts = input.split("\\^");
         String description = parts.length > 0 ? parts[0].trim() : "";
-        String from = parts.length > 1 ? parts[1].trim() : "";
-        String to = parts.length > 2 ? parts[2].trim() : "";
+        String deadline = parts.length > 1 ? parts[1].trim() : "";
 
         try {
-            return new Event(description, from, to);
+            return new Deadline(description, deadline);
         } catch (InvalidArgumentException e) {
-            throw new IllegalDataException("Failed to load event data");
+            throw new IllegalDataException("Failed to load deadline data");
         }
     }
 }
