+++ src/main/java/sylvie/storage/parser/EventParser.java
@@ -1,21 +1,22 @@
-package sylvie.storage;
+package sylvie.storage.parser;
 
 import sylvie.exception.IllegalDataException;
 import sylvie.exception.InvalidArgumentException;
-import sylvie.task.Deadline;
+import sylvie.task.Event;
 import sylvie.task.Task;
 
-public class DeadlineParser extends TaskParser {
+public class EventParser extends TaskParser {
     @Override
     public Task parse(String input) throws IllegalDataException {
         String[] parts = input.split("\\^");
         String description = parts.length > 0 ? parts[0].trim() : "";
-        String deadline = parts.length > 1 ? parts[1].trim() : "";
+        String from = parts.length > 1 ? parts[1].trim() : "";
+        String to = parts.length > 2 ? parts[2].trim() : "";
 
         try {
-            return new Deadline(description, deadline);
+            return new Event(description, from, to);
         } catch (InvalidArgumentException e) {
-            throw new IllegalDataException("Failed to load deadline data");
+            throw new IllegalDataException("Failed to load event data");
         }
     }
 }
