+++ src/main/java/sylvie/storage/parser/Parser.java
@@ -1,4 +1,4 @@
-package sylvie.storage;
+package sylvie.storage.parser;
 
 import sylvie.exception.IllegalDataException;
 import sylvie.task.Task;
