+++ src/main/java/sylvie/task/TaskList.java
@@ -4,8 +4,8 @@ import java.io.IOException;
 import java.util.ArrayList;
 import java.util.List;
 import sylvie.exception.IllegalDataException;
-import sylvie.storage.Parser;
 import sylvie.storage.Storage;
+import sylvie.storage.parser.Parser;
 import sylvie.ui.Textbox;
 
 public class TaskList {
